# Split loops into cases