This PR reduces the number of loops in row/col broadcasts in case bandwidths are preserved, which should provide an `O(n)` improvement for left/right multiplication by a `Diagonal`.

With arrays
```julia
julia> A = BandedMatrix(0=>Float64[1:3000;]); v = rand(size(A,2)); C = copy(A);
```

| Operation | Master | PR|
| --- | :-: | :-: |
| `A .* v` | 50.576 μs | 17.094 μs |
| `A  .* v'` | 51.985 μs | 19.867 μs |
| `v .* A` | 49.867 μs | 16.797 μs |
| `v' .* A` | 58.450 μs | 27.805 μs |
| `C' .= A'` | 26.480 μs | 17.158 μs |
| `C' .= A' .* 2` | 24.535 μs | 17.996 μs |
| `C' .= 2 .* A'` | 24.504 μs |17.972 μs |
| `C .= 2 .* A` | 8.767 μs | 7.899 μs |
| `C .= A .* 2` | 8.397 μs | 7.962 μs |

The performance is largely unaffected if `A` has many bands, in which case the gains from such loop splitting are small compared to the actual computation.